Fix bug in socket-controller.getMeanReadings()

+++ src/config/influx.js
@@ -14,24 +14,11 @@ import influx from 'influx'
 export let client = undefined
 
 /**
- * Establishes a connection to a database.
- *
- * @returns {Promise} Resolves to this if connection succeeded.
- */
+* Establishes a connection to a database.
+*
+* @returns {Promise} Resolves to this if connection succeeded.
+*/
 export const connectDB = async () => {
-  // Bind connection to events (to get notifications).
-  //mongoose.connection.on('connected', () => console.log('Mongoose connection is open.'))
-  //mongoose.connection.on('error', err => console.error(`Mongoose connection error has occurred: ${err}`))
-  //mongoose.connection.on('disconnected', () => console.log('Mongoose connection is disconnected.'))
-
-  // If the Node process ends, close the Mongoose connection.
-  process.on('SIGINT', () => {
-    /*mongoose.connection.close(() => {
-      console.log('Mongoose connection is disconnected due to application termination.')
-      process.exit(0)
-    })*/
-  })
-
   // Connect to the server.
   client = new influx.InfluxDB({
     database: 'readings_db',
@@ -53,13 +40,9 @@ export const connectDB = async () => {
 
   const databaseNames = await client.getDatabaseNames()
 
-  console.log(databaseNames)
-
   let containsReadingsDB = databaseNames.includes('readings_db');
 
-  console.log(containsReadingsDB)
-
-  if (containsReadingsDB) {
+  if (containsReadingsDB && process.env.RESET_DB === 'true') {
     await client.dropDatabase('readings_db')
     containsReadingsDB = false
   }
@@ -76,15 +59,15 @@ export const connectDB = async () => {
       }
     )
 
-    const addTestData = true
-
     function getRndInteger(min, max) {
       return Math.floor(Math.random() * (max - min) ) + min;
     }
 
+    const addTestData = true
+
     if (addTestData) {
       for (let i = 1; i < 45; i++) {
-        for (let j = 0; j < 24; j++) {
+        for (let j = 0; j < 5; j++) {
           let now = new Date()
           now.setDate(now.getDate() - i)
           now.setHours(j)
@@ -100,18 +83,13 @@ export const connectDB = async () => {
       }
     }
 
-    /*await client.writePoints([
-      {
-        measurement: 'readings',
-        tags: {},
-        fields: { temperature: 0, humidity: 0 }
-      }
-    ])*/
+    const results = await client.query(`
+    select * from readings
+    order by time asc
+  `)
 
-    console.log(await client.getMeasurements())
-
+    console.log(results)
   }
 
   return client
 }
-
